Killing warnings

+++ stan/math/prim/arr/functor/coupled_ode_system.hpp
@@ -124,7 +124,7 @@ namespace stan {
        */
       std::vector<double> initial_state() {
         std::vector<double> state(size_, 0.0);
-        for (int n = 0; n < N_; n++)
+        for (size_t n = 0; n < N_; n++)
           state[n] = y0_dbl_[n];
         return state;
       }
